Change name (RTTE = PathFinEx) - Refactor code (SkiExec - PathFindEx) - Unit test (A revoir quand exec implemente)
