+++ UltimateStrat/Executor/SkillExecutor.py
@@ -1,4 +1,6 @@
 #Under MIT License, see LICENSE.txt
+# TODO: Execeptions personnalisees pour les edges cases (ex: current_skill == None)
+
 from UltimateStrat.Executor.Executor import Executor
 
 __author__ = 'RoboCupULaval'
@@ -6,36 +8,38 @@ __author__ = 'RoboCupULaval'
 
 class SkillExecutor(Executor):
     """
-    SkillExecutor is a sequence of request that select next pose for each players
-    1 - what's player skill ?
-    2 - what's player target ?
-    3 - what's player goal ?
-    4 - get skill object
-    5 - generate next pose
-    6 - set next pose
+    SkillExecutor est une sequence de requetes qui choisie la prochaine
+    pose pour chaque joueur.
+
+    1 - Quelle est l'action du joueur?
+    2 - Quelle est la cible du joueur?
+    3 - Quel est le but du joueur ?
+    4 - Avoir l'objet action
+    5 - Generer la pose suivante
+    6 - Mettre en place la pose suivante
     """
     def __init__(self, info_manager):
         Executor.__init__(self, info_manager)
 
     def exec(self):
-        # Execution for each players
+        # Execution pour chaque joueur
         for id_player in range(self.info_manager.getCountPlayer()):
-            # 1 - what's player skill ?
+            # 1 - Quelle est l'action du joueur ?
             current_skill = self.info_manager.getPlayerSkill(id_player)
 
-            # 2 - what's player target ?
+            # 2 - Quelle est la cible du joueur ?
             current_target = self.info_manager.getPlayerTarget(id_player)
 
-            # 3 - what's player goal ?
+            # 3 - Quel est le but du joueur ?
             current_goal = self.info_manager.getPlayerGoal(id_player)
 
-            # 4 - get skill object
+            # 4 - Avoir l'objet action
             skill = self.skill_book[current_skill]
             
             current_pose = self.info_manager.getPlayerPose(id_player)
-            # 5 - generate next pose
+            # 5 - Generer la pose suivante
             next_pose = skill().act(current_pose, current_target, current_goal)
 
-            # 6 - set next pose
+            # 6 - Mettre en place la pose suivante
             if next_pose is not None:
                 self.info_manager.setPlayerNextAction(id_player, next_pose)
